requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
django-ckeditor==6.5.1
django-cuser==2017.3.16
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-tinymce==3.6.1
idna==3.4
Pillow==9.5.0
psycopg==3.1.9
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==1.26.16
drf-spectacular==0.26.2